+++ hera_cal/io.py
@@ -1470,14 +1470,14 @@ def write_vis(fname, data, lst_array, freq_array, antpos, time_array=None, flags
     # resort time and baseline axes
     data_array = data_array.reshape(Nblts, 1, Nfreqs, Npols)
     if nsamples is None:
-        nsample_array = np.ones_like(data_array, np.float)
+        nsample_array = np.ones_like(data_array, float)
     else:
         nsample_array = np.moveaxis(list(map(lambda p: list(map(lambda ap: nsamples[str(p)][ap], antpairs)), pols)), 0, -1)
         nsample_array = nsample_array.reshape(Nblts, 1, Nfreqs, Npols)
 
     # flags
     if flags is None:
-        flag_array = np.zeros_like(data_array, np.float).astype(np.bool)
+        flag_array = np.zeros_like(data_array, float).astype(np.bool)
     else:
         flag_array = np.moveaxis(list(map(lambda p: list(map(lambda ap: flags[str(p)][ap].astype(np.bool), antpairs)), pols)), 0, -1)
         flag_array = flag_array.reshape(Nblts, 1, Nfreqs, Npols)
@@ -1750,16 +1750,16 @@ def write_cal(fname, gains, freqs, times, flags=None, quality=None, total_qual=N
     Njones = len(jones_array)
 
     # get time info
-    time_array = np.array(times, np.float)
+    time_array = np.array(times, float)
     Ntimes = len(time_array)
-    time_range = np.array([time_array.min(), time_array.max()], np.float)
+    time_range = np.array([time_array.min(), time_array.max()], float)
     if len(time_array) > 1:
         integration_time = np.median(np.diff(time_array)) * 24. * 3600.
     else:
         integration_time = 0.0
 
     # get frequency info
-    freq_array = np.array(freqs, np.float)
+    freq_array = np.array(freqs, float)
     Nfreqs = len(freq_array)
     Nspws = 1
     freq_array = freq_array[None, :]
@@ -1769,8 +1769,8 @@ def write_cal(fname, gains, freqs, times, flags=None, quality=None, total_qual=N
     # form gain, flags and qualities
     gain_array = np.empty((Nants_data, Nspws, Nfreqs, Ntimes, Njones), np.complex)
     flag_array = np.empty((Nants_data, Nspws, Nfreqs, Ntimes, Njones), np.bool)
-    quality_array = np.empty((Nants_data, Nspws, Nfreqs, Ntimes, Njones), np.float)
-    total_quality_array = np.empty((Nspws, Nfreqs, Ntimes, Njones), np.float)
+    quality_array = np.empty((Nants_data, Nspws, Nfreqs, Ntimes, Njones), float)
+    total_quality_array = np.empty((Nspws, Nfreqs, Ntimes, Njones), float)
     for i, p in enumerate(pol_array):
         if total_qual is not None:
             total_quality_array[0, :, :, i] = total_qual[p].T[None, :, :]
@@ -1785,11 +1785,11 @@ def write_cal(fname, gains, freqs, times, flags=None, quality=None, total_qual=N
                 if quality is not None:
                     quality_array[j, :, :, :, i] = quality[(a, p)].T[None, :, :]
                 else:
-                    quality_array[j, :, :, :, i] = np.ones((Nspws, Nfreqs, Ntimes), np.float)
+                    quality_array[j, :, :, :, i] = np.ones((Nspws, Nfreqs, Ntimes), float)
             else:
                 gain_array[j, :, :, :, i] = np.ones((Nspws, Nfreqs, Ntimes), np.complex)
                 flag_array[j, :, :, :, i] = np.ones((Nspws, Nfreqs, Ntimes), np.bool)
-                quality_array[j, :, :, :, i] = np.ones((Nspws, Nfreqs, Ntimes), np.float)
+                quality_array[j, :, :, :, i] = np.ones((Nspws, Nfreqs, Ntimes), float)
 
     if total_qual is None:
         total_quality_array = None
